metrics
Macro describe_histogram

macro_rules! describe_histogram {
    ($name:expr, $unit:expr, $description:expr $(,)?) => { ... };
    ($name:expr, $description:expr $(,)?) => { ... };
}

Describes a histogram.

Histograms measure the distribution of values for a given set of measurements, and start with no initial values.

Metrics can be described with a free-form string, and optionally, a unit can be provided to describe the value and/or rate of the metric measurements. Whether or not the installed recorder does anything with the description, or optional unit, is implementation defined.

Metric names are shown below using string literals, but they can also be owned String values, which includes using macros such as format! directly at the callsite. String literals are preferred for performance where possible.

Example

// A basic histogram:
describe_histogram!("some_metric_name", "my favorite histogram");

// Providing a unit for a histogram:
describe_histogram!("some_metric_name", Unit::Bytes, "my favorite histogram");

// As mentioned in the documentation, metric names also can be owned strings, including ones
// generated at the callsite via things like `format!`:
let name = String::from("some_owned_metric_name");
describe_histogram!(name, "my favorite histogram");

describe_histogram!(format!("{}_via_format", "name"), "my favorite histogram");