metrics

Struct ProgressiveWebMetrics

pub struct ProgressiveWebMetrics {
    pub(crate) frame_type: TimerMetadataFrameType,
    pub(crate) navigation_start: Option<CrossProcessInstant>,
    pub(crate) dom_content_loaded: Cell<Option<CrossProcessInstant>>,
    pub(crate) main_thread_available: Cell<Option<CrossProcessInstant>>,
    pub(crate) time_to_interactive: Cell<Option<CrossProcessInstant>>,
    pub(crate) first_paint: Cell<Option<CrossProcessInstant>>,
    pub(crate) first_contentful_paint: Cell<Option<CrossProcessInstant>>,
    pub(crate) largest_contentful_paint: Cell<Option<CrossProcessInstant>>,
    pub(crate) time_profiler_chan: ProfilerChan,
    pub(crate) url: ServoUrl,
}

A data structure to track web metrics dfined in various specifications:

https://w3c.github.io/paint-timing/

https://github.com/WICG/time-to-interactive / https://github.com/GoogleChrome/lighthouse/issues/27

We can look at three different metrics here:

navigation start -> visually ready (dom content loaded)

navigation start -> thread ready (main thread available)

visually ready -> thread ready

Fields
frame_type: TimerMetadataFrameType

Whether or not this metric is for an <iframe> or a top level frame.

navigation_start: Option<CrossProcessInstant>

when we navigated to the page

dom_content_loaded: Cell<Option<CrossProcessInstant>>

indicates if the page is visually ready

main_thread_available: Cell<Option<CrossProcessInstant>>

main thread is available – there’s been a 10s window with no tasks longer than 50ms

time_to_interactive: Cell<Option<CrossProcessInstant>>first_paint: Cell<Option<CrossProcessInstant>>

The first paint of a particular document. TODO(mrobinson): It’s unclear if this particular metric is reflected in the specification.

See https://w3c.github.io/paint-timing/#sec-reporting-paint-timing.

first_contentful_paint: Cell<Option<CrossProcessInstant>>

The first “contentful” paint of a particular document.

See https://w3c.github.io/paint-timing/#first-contentful-paint

largest_contentful_paint: Cell<Option<CrossProcessInstant>>

The time at which the largest contentful paint was rendered.

See https://www.w3.org/TR/largest-contentful-paint/

time_profiler_chan: ProfilerChanurl: ServoUrl
Implementations

impl ProgressiveWebMetrics

pub fn new( time_profiler_chan: ProfilerChan, url: ServoUrl, frame_type: TimerMetadataFrameType, ) -> ProgressiveWebMetrics

pub(crate) fn make_metadata(&self, first_reflow: bool) -> TimerMetadata

pub fn set_dom_content_loaded(&self)

pub fn set_main_thread_available(&self, time: CrossProcessInstant)

pub fn dom_content_loaded(&self) -> Option<CrossProcessInstant>

pub fn first_paint(&self) -> Option<CrossProcessInstant>

pub fn first_contentful_paint(&self) -> Option<CrossProcessInstant>

pub fn largest_contentful_paint(&self) -> Option<CrossProcessInstant>

pub fn main_thread_available(&self) -> Option<CrossProcessInstant>

pub fn set_performance_paint_metric( &self, paint_time: CrossProcessInstant, first_reflow: bool, metric_type: ProgressiveWebMetricType, )

pub(crate) fn set_first_paint( &self, paint_time: CrossProcessInstant, first_reflow: bool, )

pub(crate) fn set_first_contentful_paint( &self, paint_time: CrossProcessInstant, first_reflow: bool, )

pub fn set_largest_contentful_paint( &self, id: LCPCandidateID, paint_time: CrossProcessInstant, area: usize, )

pub fn maybe_set_tti(&self, metric: InteractiveFlag)

pub fn get_tti(&self) -> Option<CrossProcessInstant>

pub fn needs_tti(&self) -> bool

pub fn navigation_start(&self) -> Option<CrossProcessInstant>

pub fn set_navigation_start(&mut self, time: CrossProcessInstant)

pub fn time_profiler_chan(&self) -> &ProfilerChan