metrics

Macro describe_gauge

describe_gauge!() { /* proc-macro */ }

Describes a gauge.

Gauges represent a single value that can go up or down over time, and always starts out with an initial value of zero.

Metrics can be described with a free-form string, and optionally, a unit can be provided to describe the value and/or rate of the metric measurements. Whether or not the installed recorder does anything with the description, or optional unit, is implementation defined.

Metric names are shown below using string literals, but they can also be owned String values, which includes using macros such as format! directly at the callsite. String literals are preferred for performance where possible.

Example

// A basic gauge:
describe_gauge!("some_metric_name", "my favorite gauge");

// Providing a unit for a gauge:
describe_gauge!("some_metric_name", Unit::Bytes, "my favorite gauge");

// As mentioned in the documentation, metric names also can be owned strings, including ones
// generated at the callsite via things like `format!`:
let name = String::from("some_owned_metric_name");
describe_gauge!(name, "my favorite gauge");

describe_gauge!(format!("{}_via_format", "name"), "my favorite gauge");