Trait metrics::IntoF64

pub trait IntoF64 {
    // Required method
    fn into_f64(self) -> f64;
}

An object which can be converted into a f64 representation.

This trait provides a mechanism for existing types, which have a natural representation as a 64-bit floating-point number, to be transparently passed in when recording a histogram.

Required Methods

fn into_f64(self) -> f64

Converts this object to its f64 representation.