metrics

Trait IntoLabels

pub trait IntoLabels {
    // Required method
    fn into_labels(self) -> Vec<Label>;
}

A value that can be converted to a vector of Labels.

Required Methods

fn into_labels(self) -> Vec<Label>

Consumes this value, turning it into a vector of Labels.